# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/favorite/me/{productId}": {
      "post": {
        "operationId": "FavoriteController_add",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "favorite"
        ]
      },
      "delete": {
        "operationId": "FavoriteController_remove",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "favorite"
        ]
      }
    },
...
    "/group/miniapp/me/ledger": {
      "get": {
        "operationId": "GroupMiniappController_myGroupLedger",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "我的集团余额明细（最近 N 条）",
        "tags": [
          "MiniappGroup"
        ]
      }
    },
...
    "/group/miniapp/me/card/{cardId}": {
      "get": {
        "operationId": "GroupMiniappController_myGroupCardDetail",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "我的集团洗车卡详情",
        "tags": [
          "MiniappGroup"
        ]
      }
    },
    "/group/miniapp/me/card/{cardId}/logs": {
      "get": {
        "operationId": "GroupMiniappController_myGroupCardLogs",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "我的集团洗车卡日志",
        "tags": [
          "MiniappGroup"
        ]
      }
    },
...
    "/group/miniapp/me/recharge": {
      "post": {
        "operationId": "GroupMiniappController_myGroupRecharge",
        "parameters": [
          {
            "name": "amount",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberIdForPayment",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "发起集团余额充值订单（FK），返回订单号",
        "tags": [
          "MiniappGroup"
        ]
      }
    },
    "/group/miniapp/me/members": {
      "get": {
        "operationId": "GroupMiniappController_myGroupMembers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "我的集团成员列表（管理员可见更多操作）",
        "tags": [
          "MiniappGroup"
        ]
      },
      "post": {
        "operationId": "GroupMiniappController_addMemberByPhone",
        "parameters": [
          {
            "name": "phone",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "按手机号添加集团成员（管理员）",
        "tags": [
          "MiniappGroup"
        ]
      }
    },
    "/group/miniapp/me/lookup-member-by-phone": {
      "get": {
        "operationId": "GroupMiniappController_lookupMemberByPhone",
        "parameters": [
          {
            "name": "phone",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "根据手机号查找会员（添加前预览）",
        "tags": [
          "MiniappGroup"
        ]
      }
    },
    "/group/miniapp/me/members/remove": {
      "post": {
        "operationId": "GroupMiniappController_removeMember",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "移除集团成员（管理员）",
        "tags": [
          "MiniappGroup"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wash Club API",
    "description": "API for 一体化洗车门店管理平台",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}